refactor(livewire): colocate form objects with their feature folders
Dissolve app/Livewire/Forms and move each form object next to the
components that use it. Single-form features rename their class to
Form (e.g. App\Livewire\DatabaseServer\Form); BackupForm, the
Connection rules, and NotificationChannelForm keep their names and
move under DatabaseServer/ and Configuration/ respectively.
