+++ docs/development/extending.md
@@ -17,8 +17,8 @@ All database types implement `DatabaseInterface` and are resolved via `DatabaseP
 - `app/Services/Backup/Databases/{Type}Database.php` - Create handler implementing `DatabaseInterface` (`setConfig`, `dump`, `restore`, `prepareForRestore`, `listDatabases`, `testConnection`)
 - `app/Services/Backup/Databases/DatabaseProvider.php` - Add case to `make()` and config handling in `makeForServer()`
 - `app/Services/Backup/BackupJobFactory.php` - Add snapshot creation logic if different from default (e.g., instance-level types like Redis/SQLite)
-- `app/Livewire/Forms/Connection/ConnectionRules.php` - Wire the enum case in `for()`: reuse `ClientServerConnectionRules` for host/port/credential types, or add a `{Type}ConnectionRules` subclass for type-specific rules, extra_config, dump-preview config, or field defaults
-- `app/Livewire/Forms/DatabaseServerForm.php` - Shared form behavior only (type helpers, UI behavior); per-type validation lives in the connection rules classes
+- `app/Livewire/DatabaseServer/Connection/ConnectionRules.php` - Wire the enum case in `for()`: reuse `ClientServerConnectionRules` for host/port/credential types, or add a `{Type}ConnectionRules` subclass for type-specific rules, extra_config, dump-preview config, or field defaults
+- `app/Livewire/DatabaseServer/Form.php` - Shared form behavior only (type helpers, UI behavior); per-type validation lives in the connection rules classes
 
 **UI:**
 - `resources/views/livewire/database-server/connection/{type}.blade.php` - Connection fields partial, resolved by convention from the `DatabaseType` value; `@include` the shared `_client-server-fields` partial and add type-specific fields
@@ -56,7 +56,7 @@ The volume system uses dynamic class resolution based on the type value. Use exi
 **Core:**
 - `app/Enums/VolumeType.php` - Add enum case, update `label()`, `icon()`, `sensitiveFields()`, and the `configSummary()` match. All four `match ($this)` expressions have no `default` arm, so a missing case raises an `UnhandledMatchError` at runtime (and PHPStan flags it) — a useful checklist.
 - `app/Livewire/Volume/Connectors/{Type}Config.php` - Create class extending `BaseConfig`. `BaseConfig` is a thin abstract with only two static methods to implement: `defaultConfig()` (the initial config array) and `rules(string $prefix)` (validation, keyed `{$prefix}.{field}`, using `required_if:type,{value}` for required fields).
-- `app/Livewire/Forms/VolumeForm.php` - Add a `public array ${type}Config = [];` property. **Required** — the constructor seeds every type's config from `defaultConfig()` by looping over `VolumeType::cases()`, but Livewire only binds a property that is explicitly declared. Class resolution is dynamic; the property declaration is not.
+- `app/Livewire/Volume/Form.php` - Add a `public array ${type}Config = [];` property. **Required** — the constructor seeds every type's config from `defaultConfig()` by looping over `VolumeType::cases()`, but Livewire only binds a property that is explicitly declared. Class resolution is dynamic; the property declaration is not.
 - `resources/views/livewire/volume/connectors/{type}-config.blade.php` - Create the form view. Resolved by convention from `{$form->type}-config` in `_form.blade.php` (there is no `viewName()`). Available vars: `$configPrefix`, `$readonly`, `$isEditing`. Use `<x-password>` with `:placeholder="$isEditing ? __('Leave blank to keep current') : ''"` for sensitive fields.
 - `app/Services/Backup/Filesystems/{Type}Filesystem.php` - Create class implementing `FilesystemInterface` (`handles(?string $type)` + `get(array $config)` returning a Flysystem `Filesystem`). Support both `root` (config/backup.php) and `prefix` (Volume DB) keys for the path prefix.
 - `app/Providers/AppServiceProvider.php` - Register the filesystem via `$provider->add(new {Type}Filesystem)` in the `FilesystemProvider` singleton.
@@ -76,10 +76,10 @@ The volume system uses dynamic class resolution based on the type value. Use exi
 
 ### Architecture Notes
 
-- **Dynamic Resolution**: `VolumeType::configPropertyName()` returns `{type}Config` and `configClass()` resolves `{Type}Config` from the enum value via `ucfirst()` — no explicit class mappings needed. The one non-dynamic piece is the matching public property on `VolumeForm` (above).
+- **Dynamic Resolution**: `VolumeType::configPropertyName()` returns `{type}Config` and `configClass()` resolves `{Type}Config` from the enum value via `ucfirst()` — no explicit class mappings needed. The one non-dynamic piece is the matching public property on `Volume\Form` (above).
 - **Sensitive Fields**: Fields in `sensitiveFields()` are automatically encrypted in the database, masked before browser serialization, and made optional on edit (blank = keep existing). Azure's `account_key` is an example.
 - **Connection Testing**: Works automatically via `FilesystemProvider`/`VolumeConnectionTester` (writes, reads back, deletes a temp file) once the filesystem implements `FilesystemInterface`.
-- **BaseConfig**: A minimal abstract declaring only `defaultConfig()` and `rules()`. It does not handle mounting or rendering; those live in `VolumeForm` and `_form.blade.php`.
+- **BaseConfig**: A minimal abstract declaring only `defaultConfig()` and `rules()`. It does not handle mounting or rendering; those live in `Volume\Form` and `_form.blade.php`.
 
 ---
 
@@ -99,7 +99,7 @@ The notification system uses a delegation pattern: concrete notifications extend
 - `app/Notifications/Channels/{Channel}Channel.php` - Create class with `send()` method
 
 **Configuration UI:**
-- `app/Livewire/Forms/ConfigurationForm.php` - Add properties, load/save/rules logic
+- `app/Livewire/Configuration/Form.php` - Add properties, load/save/rules logic
 - `app/Livewire/Configuration/Index.php` - Add to `getChannelOptions()`
 - `resources/views/livewire/configuration/index.blade.php` - Add conditional field section
